Temporarily disable all angular2-jwt code
We are using an extremely old version of the angular2-jwt library
(0.2.3, the current version is 5!). It has since gone through 2 major
API changes. We can't move directly to v5 as that is only supported in
Angular 9. Insteady of implementing a temporary solution, we're
commenting out auth temporarily to bring it back once we reach Angular 9
or later.